ca: make precertificate logging consistent
Previously, we logged csr.DNSNames along with the precertificate issuance
events, but this was potentially misleading. For instance, if a CSR contained
only a CN and no SANs, this set would be empty.

Instead, log the uniquified, lowercased, sorted SANs that are actually issued.

Also, emit precert=[] in issuePrecertificateInner, which is consistent with the
field name used when the final certificate is issued.