ca: log cert signing using JSON objects
This makes the log events easier to parse, and makes it easier to consistently
use the correct fields from the issuance request.

Also, reduce the number of fields that are logged on error events. Logging just
the serial and the error in most cases should suffice to cross-reference the
error with the item that we attempted to sign.