ratelimits: always use a pointer for limit
The zero value for `limit` is invalid, so returning `nil` in error cases avoids
silently returning invalid limits (and means that if code makes a mistake and
references an invalid limit it will be an obvious clear stack trace).

This is more consistent, since the methods on `limit` use a pointer
receiver. Also, since `limit` is a fairly large object, this saves some copying.