va: log MPIC summaries
Add "Perspective" and "RIR" configuration at the primary VA for each of its
backends. This information is passed through to the VA as part of the `RemoteVA`
struct.

Remove the "Hostname" field from the `RemoteVA` struct. It was previously used
in logging errors, but Perspective will be a clearer way to indicate those
errors.

Hoist the logic for "am I primary?" from `performRemoteValidation` to
PerformValidation. This allows `performRemoteValidation` to more consistently
return a set of succeeded and failed results.