Revert "wfe: on rate limit error, serve 500 (#7796)"
This reverts commit 242d74604034a0204c00ebdf3188c6eac53d508f.

We want to make this change, but it carries some risk that we'd prefer not to
take over the holiday. We'd also like to keep `main` in a state where it would
be reasonable to deploy (even if, in practice, any over-the-holiday deploy would
be a hotfix, not a direct tag from `main`).