Remove symlink from builds
This explicitly lists out the four binaries we include in releases, and drops
the symlinks and link.sh.

I think explicitly listing them is fine given that we rarely add them. This
also avoids needing to duplicate mentioning the special ct-test-srv in the
deb/tar rules.

We haven't used the symlinked versions of commands for a while, and can drop
them from builds.