grpc: factor out setup func
This uses a pattern that is new to our tests. setup accepts a variadic list of
options, and uses a type switch to make use of those options during setup. This
allows us to pass setup only the options that are relevant to any given test
case, leaving the rest to sensible defaults.