sa: GetRevokedCerts returns explicit shards too
Change GetRevokedCerts to return a combined list of certs for a given shard,
calculating shard membership temporally _and_ by explicit assignment to a shard
in the revokedCertificates table.

This functionality is gated on the ShardIdx field of GetRevokedCertsRequest.
If it is zero, revoked certs will only be returned from a given temporal shard
(and we assume that no certs have been assigned to any explicit shard yet).

After we start sending the ShardIdx field, and also start writing entries to the
revokedCertificates table, this will result in CRL sizes doubling for
several months until we retire the temporal sharding code, since most revoked
certificates will be included in one shard based on their entry in
revokedCertificates, and a different shard based on their issuance time.