Switch to loglist3 package for parsing CT log list
The schema tool used to parse log_list_schema.json doesn't work well with the
updated schema. This is going to be required to support static-ct-api logs from
current Chrome log lists.

Instead, use the loglist3 package inside the certificate-transparency-go
project, which Boulder already uses for CT submission otherwise.

As well, the Log IDs and keys returned from loglist3 have already been base64
decoded, so this re-encodes them to minimize the impact on the rest of the
codebase and keep this change small.

The test log_list.json file needed to be made a bit more realistic for loglist3
to parse without base64 or date parsing errors.