issuance: add new IncludeCRLDistributionPoints bool
To achieve this without breaking hashes of deployed configs,
create a ProfileConfigNew containing the new field (and removing
some deprecated fields).

Move the CA's profile-hashing logic into the `issuance` package,
and gate it on the presence of IncludeCRLDistributionPoints. If
that field is false (the default), create an instance of the old
`ProfileConfig` with the appropriate values and encode/hash that
instead.